lm32: eliminate truncation warnings and make EBA_RESET a parameter
CFG_EBA_RESET is supposed to be the default value of the LM32's eba_reset
parameter. Unfortunately, the constant leaked into the sources directly.

In order to clean-up a truncation warning for pc_f, it was easiest to fix
the missing parameter and then use a subarray of that value.