sdb: fix nibble order in commit_id field of SDB synthesis record. Closes #1399.
This was due to a bug in the f_string2bits() which took the commit ID as a string argument, assuming that the character array of the string is (32 downto 1), while in fact the commit_id is a string(1 to 32).

f_string2bits() is now  more resilient, it works with both "up" and "downto" string arguments, by checking the "ascending" attribute of the string type.