sw:spi: fix update spi_message->actual_length
The spi-ocore driver was not updating the actual_length field in the
spi_message structure. This field could be used by SPI users to see if
all bytes have been transferred.

In particular we found the problem because the m25p80 driver uses the
actual_length to determine if there are transmission errors (-EIO).

Signed-off-by: Mamta Shukla <mamta.ramendra.shukla@cern.ch>
Signed-off-by: Federico Vaga <federico.vaga@cern.ch>